tests: Remove conditional skipping when Typst unavailable
When this project was part of the crates.io repository, it made sense to keep the rest of the test suite running even if Typst was not installed. Since this is a dedicated project now, we can remove the conditional skipping, since anyone working on this project should have Typst installed.